pyproject.toml
[project]
name = "motionflow"
version = "0.1.0"
description = "Pose-based action detection for video streams"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0",
    "pyyaml",
    "flask",
    "opencv-python",
    "paho-mqtt",
    "OneEuroFilter",
    "numpy",
    "pynng>=0.8.1",
    "flask-sock>=0.7.0",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "config*", "utils*", "web_ui*"]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100

# Files/directories to exclude
exclude = [
    ".git",
    ".venv",
    "openvino_env",
    "__pycache__",
    "*.egg-info",
    "optimized_models",
    "yolo11*_openvino_model",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib (suggest pathlib over os.path)
    "RUF",    # Ruff-specific rules
]

# Rules to ignore
ignore = [
    "E501",   # Line too long - handled by formatter
    "B008",   # Do not perform function calls in argument defaults (pydantic uses this)
    "B017",   # Do not assert blind exception - sometimes we genuinely test for any exception
    "B027",   # Empty method in abstract class - intentional for optional hooks
    "N806",   # Variable in function should be lowercase (we use CAPS for constants)
    "N802",   # Function name should be lowercase (some test functions use different style)
    "N803",   # Argument name should be lowercase
    "ARG001", # Unused function argument (callbacks often have unused args)
    "ARG002", # Unused method argument
    "SIM102", # Collapsible if - sometimes clearer to keep separate
    "SIM108", # Use ternary operator - sometimes less readable
    "UP007",  # Use X | Y for Union - keeping Optional for clarity
    "RUF002", # Ambiguous unicode character in docstring (× is fine for math)
    "RUF003", # Ambiguous unicode character in comment
    "RUF012", # Mutable class attributes with ClassVar - Pydantic models don't need this
    "PTH",    # Pathlib suggestions - useful but not critical for now
]

# Allow autofix for all enabled rules when `--fix` is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["core", "config", "utils", "web_ui"]

[tool.ruff.lint.per-file-ignores]
# Tests can have magic values and unused imports
"tests/*" = ["PLR2004", "F401", "ARG", "B007", "RUF059"]
# Allow print and bare excepts in scripts and utils (less critical code)
"utils/*" = ["T201", "E722", "E741"]
"check_keypoints.py" = ["T201"]
"test_axelera.py" = ["T201"]
"pose_server.py" = ["T201"]

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",  # Show extra test summary info
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services or hardware",
    "system: marks end-to-end system tests with video files",
    "manual: marks tests that should only be run manually (discovery, debugging)",
]
# Timeout for individual tests (seconds)
timeout = 30

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["core", "config"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "openvino_env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"
